# final_ims/backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stock Overflow API",
    "version": "1.0.0",
    "description": "Swagger documentation generated from the implemented NestJS controllers, DTOs, and TypeScript response types."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Application API base path"
    }
  ],
...
  "paths": {
...
    "/billers": {
      "post": {
        "tags": [
          "Billers"
        ],
        "operationId": "Billers_create",
        "summary": "Create",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_BillerRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/billers_dto_create_biller_dto_CreateBillerDto"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Billers"
        ],
        "operationId": "Billers_findAll",
        "summary": "Find All",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "retailerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: retailerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "required": false,
            "description": "Query parameter: storeId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/common_database_types_BillerRecord"
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/billers/requests": {
      "post": {
        "tags": [
          "Billers"
        ],
        "operationId": "Billers_createRequest",
        "summary": "Create Request",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_BillerRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/billers_dto_create_biller_request_dto_CreateBillerRequestDto"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Billers"
        ],
        "operationId": "Billers_getRequests",
        "summary": "Get Requests",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "employeeId",
            "in": "query",
            "required": false,
            "description": "Query parameter: employeeId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/common_database_types_BillerRequest"
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/billers/requests/{id}/approve": {
      "put": {
        "tags": [
          "Billers"
        ],
        "operationId": "Billers_approveRequest",
        "summary": "Approve Request",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Path parameter: id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "biller": {
                      "$ref": "#/components/schemas/common_database_types_BillerRecord"
                    },
                    "request": {
                      "$ref": "#/components/schemas/common_database_types_BillerRequest"
                    }
                  },
                  "required": [
                    "biller",
                    "request"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/billers_dto_approve_biller_request_dto_ApproveBillerRequestDto"
              }
            }
          }
        }
      }
    },
    "/billers/requests/{id}/reject": {
      "put": {
        "tags": [
          "Billers"
        ],
        "operationId": "Billers_rejectRequest",
        "summary": "Reject Request",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Path parameter: id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_BillerRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/billers/{id}": {
      "get": {
        "tags": [
          "Billers"
        ],
        "operationId": "Billers_findOne",
        "summary": "Find One",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Path parameter: id",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_BillerRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
...
    "/returns": {
      "get": {
        "tags": [
          "Returns"
        ],
        "operationId": "Returns_findAll",
        "summary": "Find All",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "retailerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: retailerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "required": false,
            "description": "Query parameter: storeId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer",
            "in": "query",
            "required": false,
            "description": "Query parameter: customer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerEmail",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerEmail",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "supplierId",
            "in": "query",
            "required": false,
            "description": "Query parameter: supplierId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "description": "Query parameter: source",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/common_database_types_ReturnRecord"
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "Returns"
        ],
        "operationId": "Returns_create",
        "summary": "Create",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_ReturnRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/returns_dto_create_return_dto_CreateReturnDto"
              }
            }
          }
        }
      }
    },
    "/returns/{id}": {
      "get": {
        "tags": [
          "Returns"
        ],
        "operationId": "Returns_findOne",
        "summary": "Find One",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Path parameter: id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "retailerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: retailerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "required": false,
            "description": "Query parameter: storeId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer",
            "in": "query",
            "required": false,
            "description": "Query parameter: customer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerEmail",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerEmail",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "supplierId",
            "in": "query",
            "required": false,
            "description": "Query parameter: supplierId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "description": "Query parameter: source",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_ReturnRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "put": {
        "tags": [
          "Returns"
        ],
        "operationId": "Returns_update",
        "summary": "Update",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Path parameter: id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_ReturnRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/returns_dto_update_return_dto_UpdateReturnDto"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "RoleHeader": {
        "name": "role",
        "in": "header",
        "required": false,
        "description": "Caller role context for this API. Supported values include `admin`, `retailer`, `supplier`, `consumer`, and `biller`.",
        "schema": {
          "type": "string",
          "enum": [
            "admin",
            "retailer",
            "supplier",
            "consumer",
            "biller"
          ]
        }
      }
    },
    "schemas": {
...
      "billers_dto_approve_biller_request_dto_ApproveBillerRequestDto": {
        "properties": {
          "retailerId": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "billers_dto_create_biller_dto_CreateBillerDto": {
        "properties": {
          "avatar": {
            "type": "string"
          },
          "company": {
            "minLength": 1,
            "type": "string"
          },
          "country": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "phone": {
            "minLength": 1,
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          }
        },
        "required": [
          "company",
          "country",
          "email",
          "name",
          "phone"
        ],
        "type": "object"
      },
      "billers_dto_create_biller_request_dto_CreateBillerRequestDto": {
        "properties": {
          "company": {
            "minLength": 1,
            "type": "string"
          },
          "country": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "phone": {
            "minLength": 1,
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          }
        },
        "required": [
          "company",
          "country",
          "email",
          "name",
          "phone"
        ],
        "type": "object"
      },
      "common_database_types_BillerRecord": {
        "properties": {
          "avatar": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          }
        },
        "required": [
          "avatar",
          "code",
          "company",
          "country",
          "email",
          "id",
          "name",
          "phone",
          "status"
        ],
        "type": "object"
      },
      "common_database_types_BillerRequest": {
        "properties": {
          "approvedAt": {
            "type": "string"
          },
          "assignedAt": {
            "type": "string"
          },
          "assignedEmployeeId": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "rejectedAt": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "resolvedBy": {
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "type": "string"
          },
          "storeId": {
            "type": "string"
          }
        },
        "required": [
          "company",
          "country",
          "createdAt",
          "email",
          "id",
          "name",
          "phone",
          "retailerId",
          "status"
        ],
        "type": "object"
      },
...
      "common_database_types_ReturnRecord": {
        "properties": {
          "amount": {
            "type": "number"
          },
          "condition": {
            "type": "string"
          },
          "customer": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "dateN": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "emoji": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "inventoryProcessedAt": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "product": {
            "type": "string"
          },
          "productImg": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "refundMethod": {
            "type": "string"
          },
          "retailerActionAt": {
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "retailerName": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "store": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "supplierId": {
            "type": "string"
          },
          "supplierName": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "condition",
          "customer",
          "date",
          "dateN",
          "email",
          "emoji",
          "id",
          "method",
          "notes",
          "orderId",
          "priority",
          "product",
          "productImg",
          "productName",
          "reason",
          "refundMethod",
          "sku",
          "status"
        ],
        "type": "object"
      },
...
      "returns_dto_create_return_dto_CreateReturnDto": {
        "properties": {
          "amount": {
            "type": "number"
          },
          "condition": {
            "type": "string"
          },
          "customer": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emoji": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "product": {
            "type": "string"
          },
          "productImg": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "refundMethod": {
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "retailerName": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "store": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "supplierId": {
            "type": "string"
          },
          "supplierName": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "returns_dto_update_return_dto_UpdateReturnDto": {
        "properties": {
          "amount": {
            "type": "number"
          },
          "condition": {
            "type": "string"
          },
          "customer": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emoji": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "product": {
            "type": "string"
          },
          "productImg": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "refundMethod": {
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "retailerName": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "store": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "supplierId": {
            "type": "string"
          },
          "supplierName": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  }
}